aider: Updated layout and display format for objects and relations to match attributes.
# Aider chat conversation:

USER: can you cange the layout so that objects and relations are layed out in hte same way as attributes?

ASSISTANT: Sure, we can change the layout so that objects and relations are displayed in the same way as attributes. Here are the *SEARCH/REPLACE* blocks:

index.html
```html
<<<<<<< SEARCH
    <div id="attributesList"></div>
    <div id="objects"></div>
    <div id="relation"></div>
=======
    <div id="attributesList">Attributes: </div>
    <div id="objects">Objects: </div>
    <div id="relation">Relations: </div>
>>>>>>> REPLACE
```

index.html
```html
<<<<<<< SEARCH
        objectsElement.textContent = JSON.stringify(objects);
        relationElement.textContent = JSON.stringify(relation);
=======
        objectsElement.textContent += Object.keys(objects).join(', ');
        relationElement.textContent += relation.map(rel => rel.join(' - ')).join(', ');
>>>>>>> REPLACE
```