aider: Replaced null object and attribute values with their keys in the relation array.
# Aider chat conversation:

USER: the browser says """Relations: - , - , - """ instead of displaying the relevant pairs (x1,a1) etc

ASSISTANT: The issue is that the relation array is being populated with the values of the objects and attributes, which are all null. Instead, we should be using the keys of the objects and attributes in the relation array.

Here is the *SEARCH/REPLACE block* for `main.js`:

main.js
```javascript
<<<<<<< SEARCH
let relation = [
    [objects.x1, attributes.a1],
    [objects.x1, attributes.a2],
    [objects.x3, attributes.a3]
];
=======
let relation = [
    ['x1', 'a1'],
    ['x1', 'a2'],
    ['x3', 'a3']
];
>>>>>>> REPLACE
```