aider: Updated attributes object to be populated based on text content of attributesList div.
# Aider chat conversation:

USER: good, now I want to be able to change the attributes by changing the text """Attributes: a1, a2, a3, a4""" eg i want to remove "a4" from the attributes object.

ASSISTANT: To achieve this, we need to add an event listener to the `attributesList` div that will update the `attributes` object in `main.js` when the text content of the div is changed.

Here are the *SEARCH/REPLACE* blocks:

main.js
```javascript
<<<<<<< SEARCH
let attributes = {
    a1: null,
    a2: null,
    a3: null,
    a4: null
};
=======
let attributes = {};
>>>>>>> REPLACE
```

index.html
```html
<<<<<<< SEARCH
    // Function to update the data
    function updateData() {
        attributesElement.textContent += Object.keys(attributes).join(', ');
        objectsElement.textContent += Object.keys(objects).join(', ');
        relationElement.textContent += relation.map(rel => "(" + rel.join(', ') + ")").join(', ');
    }
=======
    // Function to update the data
    function updateData() {
        let attributeKeys = attributesElement.textContent.replace('Attributes: ', '').split(', ');
        attributes = attributeKeys.reduce((obj, key) => ({...obj, [key]: null}), {});
        objectsElement.textContent += Object.keys(objects).join(', ');
        relationElement.textContent += relation.map(rel => "(" + rel.join(', ') + ")").join(', ');
    }
>>>>>>> REPLACE
```

This change will make the `attributes` object in `main.js` empty by default. The `updateData` function in `index.html` will then populate the `attributes` object with keys based on the text content of the `attributesList` div.

Please note that this change will only take effect when the `updateData` function is called. If you want the `attributes` object to be updated in real time as the text content of the `attributesList` div is changed, you will need to add an event listener to the div that calls `updateData` whenever the text content changes. However, this would require making the div editable, which might not be what you want.