aider: Replaced attributes display div with input field and update button in index.html
# Aider chat conversation:

USER: ok, but the browser doesnt show a text field where i can update the new elements of attributes.

ASSISTANT: To add a text field where you can update the attributes, we need to modify the `index.html` file. We will replace the `attributesList` div with an input field and add a button that will update the attributes when clicked.

Here is the *SEARCH/REPLACE* block:

index.html
```html
<<<<<<< SEARCH
<div id="attributesList">Attributes: </div>
=======
<input id="attributesList" type="text" placeholder="Enter attributes separated by commas">
<button onclick="updateData()">Update Attributes</button>
>>>>>>> REPLACE
```

This change will replace the `attributesList` div with a text input field where you can enter the attributes. The button will call the `updateData` function when clicked, updating the attributes based on the text in the input field.