aider: Set default attributes in main.js and updated updateData function to populate input field with these defaults on page load.
# Aider chat conversation:

USER: can we do it in such a way that it loads a default from main.js?

ASSISTANT: Sure, we can modify the `main.js` file to include a default set of attributes. We can also modify the `updateData` function in the `index.html` file to populate the input field with the default attributes when the page loads.

Here are the *SEARCH/REPLACE* blocks:

main.js
```javascript
<<<<<<< SEARCH
let attributes = {};
=======
let attributes = {
    a1: null,
    a2: null,
    a3: null
};
>>>>>>> REPLACE
```

index.html
```html
<<<<<<< SEARCH
    // Function to update the data
    function updateData() {
        let attributeKeys = attributesElement.textContent.replace('Attributes: ', '').split(', ');
        attributes = attributeKeys.reduce((obj, key) => ({...obj, [key]: null}), {});
        objectsElement.textContent += Object.keys(objects).join(', ');
        relationElement.textContent += relation.map(rel => "(" + rel.join(', ') + ")").join(', ');
    }
=======
    // Function to update the data
    function updateData() {
        let attributeKeys = attributesElement.value.split(', ');
        attributes = attributeKeys.reduce((obj, key) => ({...obj, [key]: null}), {});
        objectsElement.textContent += Object.keys(objects).join(', ');
        relationElement.textContent += relation.map(rel => "(" + rel.join(', ') + ")").join(', ');
    }

    // Populate the input field with the default attributes
    attributesElement.value = Object.keys(attributes).join(', ');
>>>>>>> REPLACE
```

These changes will set a default set of attributes in the `main.js` file and populate the input field with these attributes when the page loads. The `updateData` function is also updated to read the attributes from the input field value instead of the text content.