aider: Corrected the population of relation input field with actual relation data instead of object keys.
# Aider chat conversation:

USER: currently i see the relation as """x1, x2, x3, x4""" can you correct this?

ASSISTANT: The issue is that the relation is being populated with the keys of the objects, not the actual relation data. We need to modify the `updateData` function in `index.html` to correctly populate the relation input field.

Here is the *SEARCH/REPLACE* block:

index.html
```html
<<<<<<< SEARCH
    relationElement.value = Object.keys(objects).join(', ');
=======
    relationElement.value = relation.map(rel => "(" + rel.join(', ') + ")").join(', ');
>>>>>>> REPLACE
```