aider: Fixed splitting and mapping of relation keys in updateData function.
# Aider chat conversation:

USER: """Attributes:

Objects: (x1, a1), a2), (x3, a3), (x4""" this is not correct, attributes should be: a1,a2 etc and Obuects should be x1,x2 etc

ASSISTANT: It seems like the issue is with the way the relation is being split and mapped to create the attributes and objects. The current code is splitting the relation at every comma, which is causing the attributes and objects to be incorrectly formed.

We need to modify the `updateData` function in `index.html` to correctly split the relation into attributes and objects.

Here is the *SEARCH/REPLACE* block:

index.html
```html
<<<<<<< SEARCH
    function updateData() {
        let relationKeys = relationElement.value.split(', ');
        relation = relationKeys.map(key => key.split(', '));
        let attributes = [...new Set(relation.map(rel => rel[1]))];
        let objects = [...new Set(relation.map(rel => rel[0]))];
        document.getElementById('attributesDisplay').innerText = "Attributes: " + attributes.join(', ');
        document.getElementById('objectsDisplay').innerText = "Objects: " + objects.join(', ');
    }
=======
    function updateData() {
        let relationKeys = relationElement.value.split('), ');
        relation = relationKeys.map(key => key.replace('(', '').replace(')', '').split(', '));
        let attributes = [...new Set(relation.map(rel => rel[1]))];
        let objects = [...new Set(relation.map(rel => rel[0]))];
        document.getElementById('attributesDisplay').innerText = "Attributes: " + attributes.join(', ');
        document.getElementById('objectsDisplay').innerText = "Objects: " + objects.join(', ');
    }
>>>>>>> REPLACE
```