aider: Fixed attribute and object parsing in updateData function by trimming whitespace after splitting.
# Aider chat conversation:

USER: the relation is """(x1, a1), (x1, a2), (x2,a2), (x3,a3)""" and I get """Attributes: a1, a2,

Objects: x1, x2,a2, x3,a3""" but the a3 should be among the attributes, not the objects

ASSISTANT: It seems like there might be an issue with the way the relation is being split and mapped to create the attributes and objects. The current code is splitting the relation at every comma, which is causing the attributes and objects to be incorrectly formed.

We need to modify the `updateData` function in `index.html` to correctly split the relation into attributes and objects.

Here is the *SEARCH/REPLACE* block:

index.html
```html
<<<<<<< SEARCH
    function updateData() {
        let relationKeys = relationElement.value.split('), ');
        relation = relationKeys.map(key => key.replace('(', '').replace(')', '').split(', '));
        let attributes = [...new Set(relation.map(rel => rel[1]))];
        let objects = [...new Set(relation.map(rel => rel[0]))];
        document.getElementById('attributesDisplay').innerText = "Attributes: " + attributes.join(', ');
        document.getElementById('objectsDisplay').innerText = "Objects: " + objects.join(', ');
    }
=======
    function updateData() {
        let relationKeys = relationElement.value.split('), ');
        relation = relationKeys.map(key => key.replace('(', '').replace(')', '').split(',').map(item => item.trim()));
        let attributes = [...new Set(relation.map(rel => rel[1]))];
        let objects = [...new Set(relation.map(rel => rel[0]))];
        document.getElementById('attributesDisplay').innerText = "Attributes: " + attributes.join(', ');
        document.getElementById('objectsDisplay').innerText = "Objects: " + objects.join(', ');
    }
>>>>>>> REPLACE
```
This change trims the whitespace from each attribute and object after splitting the relation, which should correct the issue.