aider: Added values to relations in main.js and adjusted calculate function to handle undefined relations.
# Aider chat conversation:

USER: the output is """DownUp: (x1, Infinity), (x2, Infinity), (x3, Infinity)""" I think "Infinity" indicates division by 0 but the current weighted upset has value 0.5 everywhere

ASSISTANT: The issue seems to be that the `relation` array in `main.js` only contains pairs of objects and attributes, but no values for the relations. This means that `relationVal[1]` in the `calculate` function in `index.html` is `undefined`, and `undefined / weight` results in `Infinity`.

To fix this, we need to add values to the relations in `main.js` and adjust the `calculate` function in `index.html` to handle the case where a relation is not defined:

Here are the *SEARCH/REPLACE* blocks:

main.js
```javascript
<<<<<<< SEARCH
let relation = [
    ['x1', 'a1'],
    ['x1', 'a2'],
    ['x2', 'a2'],
    ['x3', 'a3']];
=======
let relation = [
    ['x1', 'a1', 1],
    ['x1', 'a2', 1],
    ['x2', 'a2', 1],
    ['x3', 'a3', 1]];
>>>>>>> REPLACE
```

index.html
```html
<<<<<<< SEARCH
                let relationVal = relation.find(rel => rel[0] === object && rel[1] === attribute);
                let weight = weightedUpset.find(wu => wu[0] === attribute)[1];
                if (relationVal && weight) {
                    let val = relationVal[1] / weight;
                    if (val < minVal) {
                        minVal = val;
                    }
                }
=======
                let relationVal = relation.find(rel => rel[0] === object && rel[1] === attribute);
                let weight = weightedUpset.find(wu => wu[0] === attribute)[1];
                if (relationVal && weight) {
                    let val = relationVal[2] / weight;
                    if (val < minVal) {
                        minVal = val;
                    }
                } else if (weight) {
                    minVal = 0;
                }
>>>>>>> REPLACE
```