aider: Removed redundant attributes and objects variables from main.js
# Aider chat conversation:

USER: if I apply getObjects to the relation object, what will be the result?

ASSISTANT: The `getObjects()` method of the `Relation` class returns an array of all objects in the relation. Each object is the first element of a triple in the relation.

Given the current state of the `relation` object in `main.js`:

```javascript
let relation = new Relation([
    ['x1', 'a1', 1],
    ['x1', 'a2', 1],
    ['x2', 'a2', 1],
    ['x3', 'a3', 1]
]);
```

If you call `relation.getObjects()`, it will return `['x1', 'x1', 'x2', 'x3']`.

Please note that this is not a *SEARCH/REPLACE block* as it does not involve any changes to the code. It's just an explanation of how the `getObjects()` method works with the current state of the `relation` object.

USER: now we have the class and method `getObjects` and `getAttributes`, the lines """let attributes = ['a1', 'a2', 'a3'];

let objects = ['x1', 'x2', 'x3'];

""" in main.js are redundant, are they not?

ASSISTANT: Yes, you're correct. The `attributes` and `objects` variables in `main.js` are redundant because you can get the same information from the `relation` object using the `getAttributes()` and `getObjects()` methods.

Here is the *SEARCH/REPLACE block* to remove these lines:

main.js
```javascript
<<<<<<< SEARCH
let attributes = ['a1', 'a2', 'a3'];

let objects = ['x1', 'x2', 'x3'];
=======
>>>>>>> REPLACE
```