aider: Updated getObjects function to return a set of unique objects.
# Aider chat conversation:

USER: currently I see """Objects: x1, x1, x2, x3""" but this should be a set, so no duplicates ... should we include this in the getObjects function?

ASSISTANT: Yes, we can modify the `getObjects()` function in `main.js` to return a set of unique objects. Here is the *SEARCH/REPLACE* block:

main.js
```javascript
<<<<<<< SEARCH
getObjects() {
    return this.triples.map(triple => triple[0]);
}
=======
getObjects() {
    return [...new Set(this.triples.map(triple => triple[0]))];
}
>>>>>>> REPLACE
```