aider: Modified calculate() function to assume a default weight of 1 for attributes not specified in the weighted upset.
# Aider chat conversation:

USER: I have """(x1, a1, 1), (x1, a2, 1), (x2, a2, 1), (x3, a3, 1)""" as my relation, that is """Attributes: a1, a2, a3

Objects: x1, x2, x3
""" I changed my weighted upset to """(a1, 0.5), (a2, 0.5)""" ... in particular there is no value for a3 ... I expected now that in downUp x1 does not have value 0 anymore ... what happens if the weighted upset does not specify a value for a3? what value for a3 is assumed in the formula?

ASSISTANT: In the current implementation, if the weighted upset does not specify a value for an attribute, the weight for that attribute is considered undefined. In the `calculate()` function, if there is a relation instance for an object and an attribute, but there is no weight for that attribute (i.e., the weight is undefined), the `calculate()` function does not update `minVal` for that object. This means that if `minVal` was previously set to a number, it remains that number. If `minVal` was previously set to `Infinity`, it remains `Infinity`.

If you want the `calculate()` function to assume a default weight for attributes that are not specified in the weighted upset, you can modify the `calculate()` function to do so. For example, you could assume a default weight of 1. Here is the *SEARCH/REPLACE block* for that change:

index.html
```javascript
<<<<<<< SEARCH
                let weight = weightedUpset.find(wu => wu[0] === attribute)[1];
                if (relationInstance && weight) {
                    let val = truncatedDiv(relation.getValue(relationInstance), weight);
                    if (val < minVal) {
                        minVal = val;
                    }
                } else if (weight) {
                    minVal = 0;
                }
=======
                let weight = weightedUpset.find(wu => wu[0] === attribute);
                weight = weight ? weight[1] : 1;
                if (relationInstance) {
                    let val = truncatedDiv(relation.getValue(relationInstance), weight);
                    if (val < minVal) {
                        minVal = val;
                    }
                } else {
                    minVal = 0;
                }
>>>>>>> REPLACE
```