aider: Removed unnecessary attachment of downUp function to window object.
# Aider chat conversation:

USER: why do we need to attach it to the window object? wew dont do that for "relation" or "weightedUpset"

ASSISTANT: You're correct, attaching functions or variables to the `window` object is not always necessary. It's often done to make those functions or variables globally accessible, especially when dealing with separate script files. However, in your case, since `relation`, `weightedUpset`, and `downUp` are all defined in the same script file (`main.js`), they should all be accessible within that script without needing to be attached to the `window` object.

Let's modify the `downUp` function definition in `main.js` to not attach it to the `window` object:

main.js
```javascript
<<<<<<< SEARCH
window.downUp = function() {
=======
function downUp() {
>>>>>>> REPLACE
    let downUp = [];
    relation.getObjects().forEach(object => {
        let minVal = Infinity;
        relation.getAttributes().forEach(attribute => {
            let relationInstance = relation.triples.find(triple => relation.getObject(triple) === object && relation.getAttribute(triple) === attribute);
            let pair = weightedUpset.find(wu => wu[0] === attribute);
            let weight = pair ? pair[1] : 0; // if pair is undefined, set weight to 0
            let relationVal = relationInstance ? relation.getValue(relationInstance) : 0; // if relation undefined, set value to 0
            let val = truncatedDiv(relationVal, weight);
            if (val < minVal) {
                minVal = val;
            }
        });
        downUp.push([object, minVal]);
    });
    return downUp;
}
```