aider: Moved script tag loading main.js to the end of the body in index.html to fix 'downUp' function call before initialization error.
# Aider chat conversation:

USER: the console shows the error """Uncaught ReferenceError: can't access lexical declaration 'downUp' before initialization
    calculate file:///Users/akurz/alexhkurz-at-git/FCA/index.html:39
    onclick file:///Users/akurz/alexhkurz-at-git/FCA/index.html:1
index.html:39:22
    calculate file:///Users/akurz/alexhkurz-at-git/FCA/index.html:39
    onclick file:///Users/akurz/alexhkurz-at-git/FCA/index.html:1"""

ASSISTANT: The error message indicates that the `downUp` function is being called before it is defined. This is because the `downUp` function is defined in `main.js` but is being called in `index.html`.

To fix this, we need to ensure that `main.js` is loaded before `index.html` tries to call the `downUp` function. We can do this by moving the script tag that loads `main.js` to the end of the body in `index.html`.

Here is the *SEARCH/REPLACE* block:

index.html
```html
<<<<<<< SEARCH
<head>
    <title>Fuzzy Formal Context Analysis</title>
    <script src="main.js"></script>
</head>
<body>
=======
<body>
    <title>Fuzzy Formal Context Analysis</title>
>>>>>>> REPLACE
```

index.html
```html
<<<<<<< SEARCH
</body>
</html>
=======
    <script src="main.js"></script>
</body>
</html>
>>>>>>> REPLACE
```